+++ src/adapters/EventBusAdapterInterface.js
@@ -7,6 +7,7 @@ class EventBusAdapterInterface extends Interface
         super({
             init: 'function',
             start: 'function',
+            stop: 'function',
             subscribe: 'function',
             publish: 'function'
         });